backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Resume parsing
PyMuPDF==1.23.21
pdfplumber==0.10.3
python-docx==1.1.0

# NLP and ML
sentence-transformers==2.3.1
gensim==4.3.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Vector database
faiss-cpu==1.7.4

# LLM integration
langchain==0.1.5
langchain-community==0.0.16
langchain-openai==0.0.3
openai==1.10.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
loguru==0.7.2
python-dateutil==2.8.2
aiofiles==23.2.1

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
